add transform:translateX(100%) when the toggle show weekly or hourly
the components will slide in the screen from opposite direction when
appearing. still lookign to implement a slide out, but since the div
disappear when hourly or weekly it needs to stay on the page longer then
disappear with a transform. unfortunately this creates impagination issues for now, will need research