add spinner
the spinner handle the lack of data from the api. the
background image, searchbar and footer are loaded, then conditionally it
will check if there is data or not. if not, it will load the spinner, if
yes, it will load the data. this allows for some sort of loading screen
while the data is loaded. if no data is loaded the spinner goes on
infinitely. this solves any error screen who could appear while loading
different sets of data