fix weekly data, css adjustment
weekly data had the first item as the same day in the main graphicm but
it was assumed it was the day after, so it was showing the first item
with the day of the day after, but it was the data of today. this has
been fixede by eliminsting the first element of the data.

weekday has been added to the citydata component, as it's easier to
understand the weekdata visualization since we know what day is today

the css for weekly ad hourly data has been modified to have all the cells expand to occupy the full width. this has a better effect on small screens since it seems better formatted