add calendarday and month to WeeklyForecast, refactor calendar functions to be exportable on Services
weeklyforecast now shows weekday ("friday") as well as numerica date and
month (24th of June) to add more clarity to the information. function
Ordinal() has bene modified to allow the right value when the second
decimal is 1 2 or 3 by using the modulo inside the function. this uses a fixed amount fo pixel on screen in term of width

function Ordinal(), const MONTH and const DAYS has been moved to Services.tsx so that they dont need to be rewritten everytime they are needed

minor style change in weekly and hourly forecast