fix month in weeklyforecastitem
after creating a date variable we set the date with an increment
(date.setDate(date.getDate() + props.i)) so that at every cicle of creation of new div
for each weekday i have a date object that reflect the date of that day. this way when
i initialize and instantiate a variable "month" with the value stored, it will correctly get the month for that day.

previous implementation was statically getting the month without setting an increment
on the day object, which would have caused at the end of the month to show the same
month as today instead of the month that is suppose to be (day today is the 31st of
June, tomorrow it would have showed 1st of June instead of first of July as the Date object was set to the locale time instead of the day is supposed to show)